fix: allow wildcards in ambiguos specifiers (#41)
Sometimes metadata contains specifiers like `>=3.5.*`. This is ambiguous
because it either means `3.5.*` or `>=3.5`. The former is more
restrictive than the latter but most likely one means the latter. This
PR adds a boolean that allows a user to specify whether this should
result in an error or whether the least restrictive bound should be
chosen. With this PR the resolver chooses the least restrictive bound.