feat: extract and check wheel tags (#42)
Adds support to extract packaging tags from the Python interpreter and
use that information in the solver. This information can also be used to
determine the most compatible wheel, but although its implemented it is
not yet used in this PR.

The compatible tags are also outputted on running. Note that this could
impact the resolve result because some wheels might not be compatible
with your Python interpreter.

The compatible tags are extracted using the wonderful `packaging`
library. The library is vendored and bundled with the executable to
ensure that you still only need Python to run the rip binary.

Compatible tags are an optional argument for the resolver. If you don't
care you can just pass `None` in which case the artifacts are not
filtered based on the tag compatibility.