bump: resolvo 0.2.0
This allows indicating that some versions are excluded for reasons
outside of the solver. These reasons are included in the error message
which should make it more clear to the user why a solution could not be
found. For instance, this is especially useful if a package is only
available as an sdist.