bump: resolvo 0.2.0 (#70)
This allows indicating that some versions are excluded for reasons
outside of the solver. These reasons are included in the error message
which should make it more clear to the user why a solution could not be
found. For instance, this is especially useful if a package is only
available as an sdist.

For example when trying to resolve `sdist`.

before (`main`)

```
$ cargo run -- sdist
...
× Could not solve for the requested requirements:
│ No candidates were found for sdist *.
```

after:

```
$ cargo run -- sdist
...
× Could not solve for the requested requirements:
│ The following packages are incompatible
│ |-- sdist * cannot be installed because there are no viable options:
│     |-- sdist 0.0.0 is excluded because there are no wheels available
```