feat: create virtual environment (#82)
This is the first iteration in rip to add support for #78.

Small test was added to see if the wheel install works, this will mainly
be used for #32 but we can also integrate this into the `rip` binary to
be able to add the install path there as well. @baszalmstra let me know
if you want me to add this in a separate PR once this is merged.