fix: error in archive file matching functionality for sdists (#138)
The wrong file could be selected searching for file 'a' in archive could
also potentially find `foo/bar/a` while you were trying to find a file
in the root


Fixes #132, because a wrong `pyproject.toml` file was found for the
`ruff` source distribution, as we were using `ends_with` to match the
path of the entry in the archive. A `pyproject.toml` from a sub-crate
was selected instead of the root `pyproject.toml`.

Fixes #132, however exposes a new bug: #137.