misc: improve logging level of wheel build failures (#166)
If we find `metadata` for an artifact it does not really matter that the
others have error-ed out I think. But if we *cannot* find any then we
want to make this into an error, because currently the thread will panic
after this happens.

This should also enable pixi users to see this error.