refactor: remove all lifetimes from PackageDb and WheelBuilder to simplify code (#186)
Removes the lifetimes from `WheelBuilder` and `resolve` in favor of some
`Arc`s. I assume this will have only a very minimal impact on
performance/memory usage but it makes the code a hell of a lot easier to
understand.

The motivation to remove them is for concurrency where I often want to
`tokio::spawn` stuff. This requires that the future is `'static` which
is often not possible with all the lifetimes we used.

Most likely some lifetimes could be returned in the `resolve` function
(which don't bubble through the rest of the code) but we can do that
later.

Some of the things I turned into Arcs:

* `PackageDb`: we often want to share this between different tasks and
datastructures.
* `WheelTags`: this can amount to a lot of data so I thought it would be
worth to wrap in an Arc.
* `MarkerEnvironment`: takes up some memory, probably not worth cloning
all over the place.
* `ArtifactInfo`: This one bubbled the `'db` lifetime through to
everything. I simply used `Arc<ArtifactInfo>` now.